[AIRFLOW-6721] Move Hive2SambaOperator test
Move the Hive2SambaOperator test in the operators module into the
existing test module and use the shared test environment.
